Add inheritance of layerstree from tree
