+++ addon/templates/components/flexberry-layerstree.hbs
@@ -1,21 +1,16 @@
+{{#if _hasLayers}}
+  {{#each layers as |layer|}}
+    {{flexberry-layerstreenode
+      name=layer.name
+      type=layer.type
+      visibility=layer.visibility
+      settings=layer.settings
+      coordinateReferenceSystem=layer.coordinateReferenceSystem
+      layers=layer.layers
+      dynamicActions=layer.dynamicActions
+    }}
+  {{/each}}
+{{/if}}
 {{#if hasBlock}}
-  {{#flexberry-tree
-    class=class
-    exclusive=exclusive
-    collapsible=collapsible
-    animateChildren=animateChildren
-    duration=duration
-    nodes=nodes
-  }}
-    {{yield}}
-  {{/flexberry-tree}}
-{{else}}
-  {{flexberry-tree
-    class=class
-    exclusive=exclusive
-    collapsible=collapsible
-    animateChildren=animateChildren
-    duration=duration
-    nodes=nodes
-  }}
+  {{yield}}
 {{/if}}
