+++ addon/templates/components/flexberry-layerstreenode.hbs
@@ -1,31 +1,23 @@
-{{#if hasBlock}}
-  {{#flexberry-treenode
-    caption=name
-    iconClass=iconClass
-    hasCheckbox=true
-    checkboxValue=visibility
-    readonly=false
-    headerClick=headerClick
-    beforeExpand=beforeExpand
-    beforeCollapse=beforeCollapse
-    checkboxChange=visibilityChange
-    checkboxCheck=becameVisible
-    checkboxUncheck=becameInvisible
-  }}
+{{flexberry-treenode
+  caption=name
+  iconClass=iconClass
+  hasCheckbox=true
+  checkboxValue=visibility
+  readonly=false
+  headerClick=headerClick
+  beforeExpand=beforeExpand
+  beforeCollapse=beforeCollapse
+  checkboxChange=visibilityChange
+  checkboxCheck=becameVisible
+  checkboxUncheck=becameInvisible
+  dynamicActions=_convertedDynamicActions
+  hasNodesOuter=(or hasBlock _hasLayers)
+}}
+<div class="{{flexberryClassNames.treeNodeContent}} content">
+  {{#if _hasLayers}}
+    {{flexberry-layerstree layers=layers}}
+  {{/if}}
+  {{#if hasBlock}}
     {{yield}}
-  {{/flexberry-treenode}}
-{{else}}
-  {{flexberry-treenode
-    caption=name
-    iconClass=iconClass
-    hasCheckbox=true
-    checkboxValue=visibility
-    readonly=false
-    headerClick=headerClick
-    beforeExpand=beforeExpand
-    beforeCollapse=beforeCollapse
-    checkboxChange=visibilityChange
-    checkboxCheck=becameVisible
-    checkboxUncheck=becameInvisible
-  }}
-{{/if}}
+  {{/if}}
+</div>
