+++ addon/templates/components/flexberry-treenode.hbs
@@ -1,7 +1,7 @@
 <div
   class="{{flexberryClassNames.treeNodeHeader}} title"
-  onclick={{action "onHeaderClick" (or hasBlock _hasNodes) preventDefault=false}}>
-  {{#if (or hasBlock _hasNodes)}}
+  onclick={{action "onHeaderClick" (or hasBlock (or _hasNodes hasNodesOuter)) preventDefault=false}}>
+  {{#if (or hasBlock (or _hasNodes hasNodesOuter))}}
     <i class="{{flexberryClassNames.treeNodeExpandCollapseIcon}} dropdown icon"></i>
   {{else}}
     {{!-- Here "visibility: hidden;" style is to force exactly the same padding for tree nodes without children --}}
@@ -22,11 +22,13 @@
   {{/if}}
   {{caption}}
 </div>
-<div class="{{flexberryClassNames.treeNodeContent}} content">
-  {{#if _hasNodes}}
-    {{flexberry-tree nodes=nodes}}
-  {{/if}}
-  {{#if hasBlock}}
-    {{yield}}
-  {{/if}}
-</div>
+{{#if (not hasNodesOuter)}}
+  <div class="{{flexberryClassNames.treeNodeContent}} content">
+    {{#if _hasNodes}}
+      {{flexberry-tree nodes=nodes}}
+    {{/if}}
+    {{#if hasBlock}}
+      {{yield}}
+    {{/if}}
+  </div>
+{{/if}}
